requirements.txt
argon2-cffi
fastapi
uvicorn[standard]
passlib
python-jose
pydantic
sqlalchemy
python-dotenv
